refactor move jwt classes into provider package, make them package private
