+++ auth0/src/main/java/com/auth0/android/provider/NoSignatureVerifier.java
@@ -1,0 +1,12 @@
+package com.auth0.android.provider;
+
+class NoSignatureVerifier extends SignatureVerifier {
+
+    private static final String EXPECTED_ALGORITHM = "HS256";
+
+    NoSignatureVerifier() {
+        super(EXPECTED_ALGORITHM);
+        //TODO: anything missing?
+    }
+
+}
