+++ auth0/src/main/java/com/auth0/android/provider/TokenValidationException.java
@@ -1,12 +1,11 @@
-package com.auth0.android.jwt;
+package com.auth0.android.provider;
 
 import com.auth0.android.authentication.AuthenticationException;
 
-//TODO: Make pkg private
-public class TokenValidationException extends AuthenticationException {
+class TokenValidationException extends AuthenticationException {
     private static final String ERROR_CODE = "a0.sdk.internal_error.id_token_validation";
 
-    public TokenValidationException(String message) {
+    TokenValidationException(String message) {
         super(ERROR_CODE, message);
     }
 }
