+++ auth0/src/test/java/com/auth0/android/request/internal/UserProfileGsonTest.java
@@ -199,7 +199,6 @@ public class UserProfileGsonTest extends GsonBaseTest {
         UserProfile profile = pojoFrom(json(PROFILE_FULL), UserProfile.class);
         assertThat(profile, isNormalizedProfile(ID, NAME, NICKNAME));
         assertThat(profile.getExtraInfo(), hasEntry("multifactor", (Object) Collections.singletonList("google-authenticator")));
-        // FIXME: Weird classpath issue with hamcrest. Probably best to rewrite assets. (06/07/2016)
         assertThat(profile.getExtraInfo(), not(anyOf(new Matcher[]{hasKey("user_id"), hasKey("name"), hasKey("nickname"), hasKey("picture"), hasKey("email"), hasKey("created_at")})));
         assertThat(profile.getExtraInfo(), not(anyOf(new Matcher[]{hasKey("identities"), hasKey("user_metadata"), hasKey("app_metadata")})));
     }
